evm: split fork and integration tests
